Restructure lib folder
Loader now has a home. And the random `ghost.ts` file also has a home
